Controleur.php: controle que le formulaire a été soumis à la création et l'update d'une annonce. annonce_detail.php: controle permettant de cacher les liens de modification et suppression si l'utilisateur qui reegarde l'annonce n'est pas le propriétaire.

+++ Vue/annonce_detail.php
@@ -1,6 +1,4 @@
-<h1>Votre annonce</h1>
-
-<h2><?= htmlspecialchars($annonce['TITRE_ANNONCE'] ?? '') ?></h2>
+<h1><?= htmlspecialchars($annonce['TITRE_ANNONCE'] ?? '') ?></h1>
 
 <p><?= nl2br(htmlspecialchars($annonce['DESCRIPTION'] ?? '')) ?></p>
 
@@ -12,10 +10,13 @@
     <img src="ressources/imagesAnnonces/<?= $img['chemin_image'] ?>"
         style="max-width:200px; margin:10px;">
 <?php endforeach; ?>
-<br>
-<a href="index.php?action=editAnnonce&id=<?= $annonce['ID_ANNONCE'] ?>">Modifier l’annonce</a>
-<br>
-<a href="index.php?action=deleteAnnonce&id=<?= $annonce['ID_ANNONCE'] ?>"
-    onclick="return confirm('Voulez-vous vraiment supprimer cette annonce ?');">
-    Supprimer l’annonce
-</a>
+
+<?php if (isset($_SESSION['user']) && $_SESSION['user']['id'] == $annonce['ID_USER']): ?>
+    <br>
+    <a href="index.php?action=editAnnonce&id=<?= $annonce['ID_ANNONCE'] ?>">Modifier l'annonce</a>
+    <br>
+    <a href="index.php?action=deleteAnnonce&id=<?= $annonce['ID_ANNONCE'] ?>"
+        onclick="return confirm('Voulez-vous vraiment supprimer cette annonce ?');">
+        Supprimer l'annonce
+    </a>
+<?php endif; ?>
